fix: move runmode
