+++ crates/jstz_node/src/config.rs
@@ -1,10 +1,64 @@
-use std::path::{Path, PathBuf};
+use std::{
+    fmt::Display,
+    path::{Path, PathBuf},
+};
 
+use anyhow::Context;
 use jstz_utils::KeyPair;
 use octez::r#async::endpoint::Endpoint;
-use serde::Serialize;
+use serde::{Deserialize, Serialize};
+use tempfile::NamedTempFile;
 
-use crate::RunMode;
+#[derive(Debug, Deserialize, Serialize, Clone, PartialEq)]
+#[serde(rename_all = "lowercase")]
+#[serde(tag = "mode")]
+pub enum RunMode {
+    Sequencer {
+        capacity: usize,
+        debug_log_path: PathBuf,
+    },
+    #[serde(alias = "default")]
+    Default,
+}
+
+impl Default for RunMode {
+    fn default() -> Self {
+        Self::Default
+    }
+}
+
+impl Display for RunMode {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            RunMode::Default => write!(f, "default"),
+            RunMode::Sequencer { .. } => write!(f, "sequencer"),
+        }
+    }
+}
+
+impl RunMode {
+    pub fn new(
+        mode_str: Option<&str>,
+        capacity: Option<usize>,
+        debug_log_path: Option<PathBuf>,
+    ) -> anyhow::Result<Self> {
+        match mode_str.unwrap_or("default") {
+            "sequencer" => Ok(RunMode::Sequencer {
+                capacity: capacity.unwrap_or(1),
+                debug_log_path: debug_log_path.unwrap_or(
+                    NamedTempFile::new()
+                        .context("failed to create temporary debug log file")?
+                        .into_temp_path()
+                        .keep()
+                        .context("failed to convert temporary debug log file to path")?
+                        .to_path_buf(),
+                ),
+            }),
+            "default" => Ok(RunMode::Default),
+            v => Err(anyhow::anyhow!("invalid run mode '{v}'")),
+        }
+    }
+}
 
 #[derive(Clone, Serialize)]
 pub struct JstzNodeConfig {
@@ -113,5 +167,61 @@ mod tests {
         assert_eq!(json["mode"], "sequencer");
         assert_eq!(json["capacity"], 123);
         assert_eq!(json["debug_log_path"], "/debug/log");
+    }
+
+    #[test]
+    fn default_runmode() {
+        assert_eq!(RunMode::default(), RunMode::Default);
+    }
+
+    #[test]
+    fn runmode_to_string() {
+        assert_eq!(RunMode::Default.to_string(), "default");
+        assert_eq!(
+            RunMode::Sequencer {
+                capacity: 1,
+                debug_log_path: PathBuf::new()
+            }
+            .to_string(),
+            "sequencer"
+        );
+    }
+
+    #[test]
+    fn runmode_new() {
+        assert_eq!(RunMode::new(None, None, None).unwrap(), RunMode::Default);
+        assert_eq!(
+            RunMode::new(Some("default"), None, None).unwrap(),
+            RunMode::Default
+        );
+
+        let mode = RunMode::new(Some("sequencer"), None, None).unwrap();
+        matches!(
+            mode,
+            RunMode::Sequencer {
+                capacity: 1,
+                debug_log_path: _
+            }
+        );
+
+        assert_eq!(
+            RunMode::new(
+                Some("sequencer"),
+                Some(123),
+                Some(PathBuf::from_str("/foo/bar").unwrap())
+            )
+            .unwrap(),
+            RunMode::Sequencer {
+                capacity: 123,
+                debug_log_path: PathBuf::from_str("/foo/bar").unwrap()
+            }
+        );
+
+        assert_eq!(
+            RunMode::new(Some("bad"), None, None)
+                .unwrap_err()
+                .to_string(),
+            "invalid run mode 'bad'"
+        );
     }
 }
